requirements.txt
nose
sphinx
pyserial